+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog for package leap_motion
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.0.12 (2018-03-06)
+-------------------
+* Reimplementation of the entire Leap Motion driver for ROS
+* Old implementation of the driver is now deprecated and will be removed in a year.
+* Contributors: Nowittyusername
+
 0.0.11 (2017-01-14)
 -------------------
 * [fix] ROS Hydro onward requires queue_size option `#31 <https://github.com/ros-drivers/leap_motion/issues/31>`_